+++ packages/Webkul/DataGrid/src/Providers/DataGridServiceProvider.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Webkul\DataGrid\Providers;
+
+use Illuminate\Support\ServiceProvider;
+
+class DataGridServiceProvider extends ServiceProvider
+{
+    /**
+     * Bootstrap any application services.
+     */
+    public function boot(): void
+    {
+    }
+
+    /**
+     * Register any application services.
+     */
+    public function register(): void
+    {
+    }
+}
