+++ packages/Webkul/Ui/src/Providers/ModuleServiceProvider.php
@@ -6,6 +6,5 @@ use Webkul\Core\Providers\CoreModuleServiceProvider;
 
 class ModuleServiceProvider extends CoreModuleServiceProvider
 {
-    protected $models = [
-    ];
+    protected $models = [];
 }
