+++ packages/Webkul/Ui/src/Providers/UiServiceProvider.php
@@ -34,6 +34,5 @@ class UiServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $this->app->bind('datagrid', 'Webkul\Ui\DataGrid\DataGrid');
     }
 }
